workspaces/api/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "workspaces-api"
version = "0.0.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg",
    "httpx",
    "pydantic>=2",
    "pydantic-settings>=2",
    "alembic>=1.13",
    "sse-starlette>=2",
    "psycopg[binary]>=3",
    "pyjwt[crypto]>=2.8",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff",
    "mypy",
    "aiosqlite",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
